refactor(src): extract github/link_header.rs shared helper
